Extract custom print method to lib
